Implementação da lib SweetAlert

+++ web-app/assets/js/publicacoes.js
@@ -20,7 +20,7 @@ function criarPublicacao(evento) {
     }).done(function () {
         window.location = "/home";
     }).fail(function () {
-        alert("Erro ao criar a publicação!");
+        Swal.fire("Ops....", "Erro ao criar a publicação!", "error");
     });
 }
 
@@ -46,7 +46,7 @@ function curtirPublicacao(evento) {
         elementoClicado.addClass('text-danger');
         elementoClicado.removeClass('curtir-publicacao');
     }).fail(function () {
-        alert("Erro ao curtir a publicação");
+        Swal.fire("Ops....", "Erro ao curtir a publicação!", "error");
     }).always(function () {
         elementoClicado.prop('disabled', false);
     });
@@ -74,7 +74,7 @@ function descurtirPublicacao(evento) {
         elementoClicado.removeClass('text-danger');
         elementoClicado.addClass('curtir-publicacao');
     }).fail(function () {
-        alert("Erro ao descurtir a publicação");
+        Swal.fire("Ops....", "Erro ao descurtir a publicação!", "error");
     }).always(function () {
         elementoClicado.prop('disabled', false);
     });
@@ -93,9 +93,11 @@ function atualizarPublicacao() {
             conteudo: $('#conteudo').val(),
         }
     }).done(function () {
-        alert("Publicação editada com sucesso!")
+        Swal.fire("Sucesso", "Publicação editada com sucesso!", "success").then(function () {
+            window.location = "/home"
+        })
     }).fail(function () {
-        alert("Erro ao editar a publicação!")
+        Swal.fire("Ops....", "Erro ao editar a publicação!", "error");
     }).always(function () {
         $('#atualizar-publicacao').prop('disabled', false);
     });
@@ -104,20 +106,30 @@ function atualizarPublicacao() {
 function deletarPublicacao(evento) {
     evento.preventDefault();
 
-    const elementoClicado = $(evento.target);
-    const publicacao = elementoClicado.closest('div');
-    const publicacaoID = publicacao.data('publicacao-id');
+    Swal.fire({
+        title: "Atenção!",
+        text: "Tem certeza que deseja excluir essa publicação?",
+        showCancelButton: true,
+        cancelButtonText: "Cancelar",
+        icon: "warning"
+    }).then(function (confirmacao) {
+        if (!confirmacao.value) return;
 
-    elementoClicado.prop('disabled', true);
+        const elementoClicado = $(evento.target);
+        const publicacao = elementoClicado.closest('div');
+        const publicacaoID = publicacao.data('publicacao-id');
 
-    $.ajax({
-        url: `/publicacoes/${publicacaoID}`,
-        method: "DELETE"
-    }).done(function () {
-        publicacao.fadeOut("slow", function () {
-            $(this).remove();
+        elementoClicado.prop('disabled', true);
+
+        $.ajax({
+            url: `/publicacoes/${publicacaoID}`,
+            method: "DELETE"
+        }).done(function () {
+            publicacao.fadeOut("slow", function () {
+                $(this).remove();
+            });
+        }).fail(function () {
+            Swal.fire("Ops....", "Erro ao excluir a publicação!", "error");
         });
-    }).fail(function () {
-        alert("Erro ao excluir a publicação!")
     });
 }
